Extract express app from entry point file
